% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_plot.R
\name{labeled_plot}
\alias{labeled_plot}
\title{Attach labels to a plot object}
\usage{
labeled_plot(p, overwrite = FALSE, ...)
}
\arguments{
\item{p}{Plot to wrap}

\item{overwrite}{Overwrite existing files by default.}

\item{...}{Other parameters passed to saving method (e.g.
see \code{\link[ggplot2]{ggsave}})}

\item{filename}{Name of file to save to, including extension. Will be added
to the path when saving plot lists. Overides \code{file_format}}

\item{file_format}{Format of file to save (e.g. pdf, jpg)}

\item{units, height, width}{Size of plot, in \code{units}}
}
\description{
\code{labeled_plot} creates a plot object with attributes describing how to
save it
}
\details{
This is an S3 class interacting with the \code{smart_save} S3 generic.
}
