% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlistr.R
\docType{package}
\name{plotlistr}
\alias{plotlistr}
\alias{plotlistr-package}
\title{plotlistr: A package for managing lists of plots}
\description{
Tools to work with lists of plots, allowing you to automatically generate
plots during an analysis and save them with sensible parameters.
}
\section{plotlistr functions}{

}

