% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlist.R
\name{save_plotlist}
\alias{save_plotlist}
\title{Save lists of plots.}
\usage{
save_plotlist(
  plotlist,
  root,
  overwrite = c("none", "all", "plot"),
  verbose = 1,
  default_format = "pdf"
)
}
\arguments{
\item{plotlist}{Potentially nested list of plot objects.}

\item{root}{Root directory to save to.}

\item{overwrite}{Whether to overwrite existing files, with "plot" respecting
the property stored in a \code{\link{labeled_plot}}.}

\item{verbose}{Maximum list depth to report saving}

\item{default_format}{Default file format to save in}
}
\description{
Recursively save plots in a (nested) list, allowing you to generate all
plots during an analysis and then plot them afterwards. Use
\code{\link{labeled_plot}} to add properties so that plots can be easily
saved automatically.
}
\details{
This function uses the \code{\link{smart_save}} method to save plots,
respecting the attributes of \code{\link{labeled_plot}} objects. Nested
levels of the list will be saved in subfolders
}
