% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_save.R
\name{smart_save}
\alias{smart_save}
\alias{smart_save.default}
\alias{smart_save.labeled_plot}
\alias{smart_save.multipanelfigure}
\title{Inteligently save plot objects.}
\usage{
smart_save(p, ...)

\method{smart_save}{default}(p, filename, ...)

\method{smart_save}{labeled_plot}(p, filename = NULL, override = FALSE, ...)

\method{smart_save}{multipanelfigure}(p, filename, ...)
}
\arguments{
\item{p}{Plot object to save.}

\item{...}{Additional parameters passed to plot method.}

\item{filename}{Path to save plot at.}

\item{override}{Override labels from \code{labeled_plot} with values from
\code{...}.}
}
\description{
\code{smart_save} saves plot objects using different methods depending on
class, including support for \code{labeled_plot} objects.
}
\details{
This is a generic function with methods defined for different classes,
dispatching on the \code{p} argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Use \code{ggsave} by default

\item \code{labeled_plot}: Process labels from \code{labeled_plot}

\item \code{multipanelfigure}: Save \code{multipanelfigure} objects
}}

