% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlist.R
\name{sv_pltlst}
\alias{sv_pltlst}
\title{Internal recursive function for saving plotlists}
\usage{
sv_pltlst(p, root, overwrite, verbose, level = 1,
  default_format = "pdf")
}
\arguments{
\item{p}{Plot or list to save.}

\item{root}{Root directory.}

\item{overwrite}{Level of overwriting from \code{c("none", "all", "plot")}.}

\item{verbose}{Maximum depth to report at.}

\item{level}{Current depth.}

\item{default_format}{Default file format to save to.}
}
\description{
This function performs the recursive saving for \code{\link{save_plotlist}}.
It's parameters mostly link to those in \code{save_plotlist}.
}
\keyword{internal}
